Four files in utils with charCounter and getEpisodesLocations functions and respectives test files added.
