+++ src/pages/challenge.js
@@ -1,66 +1,12 @@
 import { fetchData } from '../services/apiFetch';
 import { useState, useEffect } from 'react';
+import charCounter from '../utils/charCounter';
+import getEpisodesLocations from '../utils/getEpisodesLocations';
 
 export default function Challenge() {
   const [characters, setCharacters] = useState([]);
   const [episodes, setEpisodes] = useState([]);
   const [locations, setLocations] = useState([]);
-
-  //counterLetter function return the number of characters than are equal to the letter passed as argument:
-
-  function counterLetters(data, letter){
-    let t0 = performance.now();
-    let counterLetter = 0;
-  
-    for(let episode in data){  //for each episode in the data
-        let name = data[episode].name;  //get the name of the episode
-        for (let i in name){  //for each letter in the name
-          counterLetter += (name[i].toLowerCase() === letter ? 1 : 0);  //if the letter is equal to the letter passed as argument, add 1 to the counterLetter
-        }
-    }
-    let t1 = performance.now();
-    let totalTimeCharCounter = JSON.stringify(t1 - t0).slice(0,1) + 's ' + JSON.stringify(t1 - t0).slice(2,10) + 'ms';
-    localStorage.setItem('timeCharCounter', totalTimeCharCounter);  //save the function execution time in the local storage
-
-    return counterLetter; 
-  };
-
-  //getEpisodesLocations function return the locations name for each episode
-
-  function getEpisodesLocations(){
-    let t0 = performance.now();
-    let locationsArray = [];  //array to store the locations name for each episode
-  
-    for (let episode in episodes){  //for each episode in episodes seted in the state
-      const episodeName = episodes[episode].name;  //get the name of the episode
-      const episodeCode = episodes[episode].episode;  //get the code of the episode
-      let idCharacters = [];  //array to store the id of the characters in the episode
-      let episodeLocations = [];  //array to store the locations name for each episode
-  
-      for (let url in episodes[episode].characters){  //for each url in the characters array of the episode
-        idCharacters.push(parseInt(url.slice(url.lastIndexOf('/') + 1)));  //slice the url to get the character's id and push it to the idCharacters array
-      }
-
-      characters.map(character => {  //for each character in the characters seted in the state
-        if(idCharacters.includes(character.id)){  //if character's id is in the idCharacters array
-          episodeLocations.push(character.origin.name);  //add the character's location name to the episodeLocations array
-        }
-      })
-
-      locationsArray.push({  //push the episode's locations name to the locationsArray array
-        "name": episodeName,
-        "episode": episodeCode,
-        "locations": episodeLocations.filter((v, i, a) => a.indexOf(v) === i) // filter for unique values in locations
-      })
-    }
-    let t1 = performance.now();
-    let totalTimeLocations = JSON.stringify(t1 - t0).slice(0,1) + 's ' + JSON.stringify(t1 - t0).slice(2,10) + 'ms';
-
-    localStorage.setItem('timeLocations', totalTimeLocations);  //save the function execution time in the local storage
-
-    return locationsArray
-  };
-
 
   useEffect(() => {
     async function fetchAllData() {
@@ -94,17 +40,17 @@ export default function Challenge() {
   "results": [
     {
       "char": "l",
-      "count": ${counterLetters(locations, 'l') || 0}
+      "count": ${charCounter(locations, 'l') || 0}
       "resource": "location"
     },
     {
       "char": "e",
-      "count": ${counterLetters(episodes, 'e') || 0}
+      "count": ${charCounter(episodes, 'e') || 0}
       "resource": "episode"
     },
     {
       "char": "c",
-      "count": ${counterLetters(characters, 'c') || 0}
+      "count": ${charCounter(characters, 'c') || 0}
       "resource": "character"
     }
   ]
@@ -113,7 +59,7 @@ export default function Challenge() {
   "exercise_name": "Episode locations",
   "time": ${localStorage.getItem('timeLocations')  || '0s 0ms'},
   "in_time": true,
-  "results": ${JSON.stringify(getEpisodesLocations(), null, 2) || '[]'}
+  "results": ${JSON.stringify(getEpisodesLocations(episodes, characters), null, 2) || '[]'}
 }
 `}
           </code>
